feature AuthOnLibrary
toTopBtn added to Library as well,
toTopBtn logic moved to separete JS file
