+++ src/index.html
@@ -25,8 +25,8 @@
     <include src="./partials/main.html"></include>
     <include src="./partials/modal-window.html"></include>
     <include src="./partials/pagination.html"></include>
-    <include src="./partials/footer.html"></include
-    ><include src="./partials/login-form.html"></include>
+    <include src="./partials/footer.html"></include>
+    <include src="./partials/login-form.html"></include>
     <include src="./partials/students.html"></include>
   </body>
 </html>
