+++ src/index.js
@@ -4,6 +4,6 @@ import { advencedSearch } from './js/advencedSearch.js';
 import { addToLocalStorage } from './js/addToLocalStorage.js';
 import { modalWindow } from './js/modalWindow.js';
 import { renderFoundMovies } from './js/searchAndRenderMovies.js';
+import './js/toTopBtn';
 import './js/students';
 import './js/firebase';
-
