+++ src/js/searchAndRenderMovies.js
@@ -6,7 +6,7 @@ const header = document.querySelector('.header__no-movies');
 const main = document.querySelector('.main-box');
 const inputSearch = document.querySelector("input[name='searchQuery']");
 const searchButton = document.querySelector('.header__search-form-btn');
-const toTopBtn = document.querySelector('.to-top');
+// const toTopBtn = document.querySelector('.to-top');
 
 let currentPage = 1;
 
@@ -81,23 +81,23 @@ async function renderFoundMovies(event) {
 
 export { renderFoundMovies };
 
-toTopBtn.addEventListener('click', onTopScroll);
+// toTopBtn.addEventListener('click', onTopScroll);
 
-window.addEventListener('scroll', onScrollToTopBtn);
+// window.addEventListener('scroll', onScrollToTopBtn);
 
-function onScrollToTopBtn() {
-  const offsetTrigger = 100;
-  const pageOffset = window.pageYOffset;
+// function onScrollToTopBtn() {
+//   const offsetTrigger = 100;
+//   const pageOffset = window.pageYOffset;
 
-  pageOffset > offsetTrigger
-    ? toTopBtn.classList.remove('is-hidden')
-    : toTopBtn.classList.add('is-hidden');
-}
+//   pageOffset > offsetTrigger
+//     ? toTopBtn.classList.remove('is-hidden')
+//     : toTopBtn.classList.add('is-hidden');
+// }
 
-function onTopScroll() {
-  window.scrollTo({
-    top: 0,
-    behavior: 'smooth',
-  });
-}
-export { onTopScroll };
+// function onTopScroll() {
+//   window.scrollTo({
+//     top: 0,
+//     behavior: 'smooth',
+//   });
+// }
+// export { onTopScroll };
