+++ src/library.html
@@ -3,15 +3,16 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Page title</title>
+    <title>Library</title>
     <link rel="stylesheet" href="./sass/library.scss" />
     <script type="module" src="./library.js"></script>
   </head>
+
   <body>
-    <header>
     <include src="./partials/myLibraryHeader.html"></include>
     <include src="./partials/main.html"></include>
     <include src="./partials/modal-window.html"></include>
+    <include src="./partials/login-form.html"></include>
     <include src="./partials/footer.html"></include>
   </body>
 </html>
