+++ src/library.js
@@ -1,3 +1,5 @@
 import './js/getGenreList';
 import './js/getMoviesById';
 import './js/modalWindow';
+import './js/firebase';
+import './js/toTopBtn'
