+++ src/sass/library.scss
@@ -4,4 +4,7 @@
 @import './components/myLibraryHeader';
 @import './components/main';
 @import './components/modal-window';
+@import './components/students';
+@import './components/login-form';
 @import './components/footer';
+@import './components/to-top-btn';
